+++ src/test/groovy/edu/ie3/datamodel/io/factory/input/Transformer3WInputFactoryTest.groovy
@@ -89,4 +89,3 @@ class Transformer3WInputFactoryTest  extends Specification implements FactoryTes
     e.cause.message == "Voltage level of node a must be greater than voltage level of node b and voltage level of node b must be greater than voltage level of node c"
   }
 }
-