+++ src/test/groovy/edu/ie3/datamodel/io/factory/input/participant/FixedFeedInInputFactoryTest.groovy
@@ -5,6 +5,8 @@
  */
 package edu.ie3.datamodel.io.factory.input.participant
 
+import static edu.ie3.util.quantities.PowerSystemUnits.PU
+
 import edu.ie3.datamodel.exceptions.FactoryException
 import edu.ie3.datamodel.io.factory.input.NodeAssetInputEntityData
 import edu.ie3.datamodel.models.StandardUnits
@@ -17,10 +19,8 @@ import edu.ie3.test.helper.FactoryTestHelper
 import spock.lang.Specification
 import tech.units.indriya.quantity.Quantities
 
-import javax.measure.quantity.Dimensionless
 import java.time.ZonedDateTime
-
-import static edu.ie3.util.quantities.PowerSystemUnits.PU
+import javax.measure.quantity.Dimensionless
 
 class FixedFeedInInputFactoryTest extends Specification implements FactoryTestHelper {
   def "A FixedFeedInInputFactory should contain exactly the expected class for parsing"() {