+++ src/test/groovy/edu/ie3/datamodel/models/input/container/SystemParticipantsTest.groovy
@@ -22,8 +22,7 @@ class SystemParticipantsTest extends Specification {
         Collections.singleton(SystemParticipantTestData.loadInput),
         Collections.singleton(SystemParticipantTestData.pvInput),
         Collections.singleton(SystemParticipantTestData.storageInput),
-        Collections.singleton(SystemParticipantTestData.wecInput),
-        Collections.singleton(SystemParticipantTestData.emInput)
+        Collections.singleton(SystemParticipantTestData.wecInput)
         )
 
     when:
@@ -45,8 +44,7 @@ class SystemParticipantsTest extends Specification {
         Collections.singleton(SystemParticipantTestData.loadInput),
         Collections.singleton(SystemParticipantTestData.pvInput),
         Collections.singleton(SystemParticipantTestData.storageInput),
-        Collections.singleton(SystemParticipantTestData.wecInput),
-        Collections.singleton(SystemParticipantTestData.emInput)
+        Collections.singleton(SystemParticipantTestData.wecInput)
         )
 
     def modifiedBmInput = SystemParticipantTestData.bmInput.copy().id("modified").build()
@@ -59,13 +57,12 @@ class SystemParticipantsTest extends Specification {
     def modifiedPvInput = SystemParticipantTestData.pvInput.copy().id("modified").build()
     def modifiedStorageInput = SystemParticipantTestData.storageInput.copy().id("modified").build()
     def modifiedWecInput = SystemParticipantTestData.wecInput.copy().id("modified").build()
-    def modifiedEmInput = SystemParticipantTestData.emInput.copy().id("modified").build()
 
     when:
     def modifiedSystemParticipants = systemParticipants.copy()
         .bmPlants(Set.of(modifiedBmInput))
         .chpPlants(Set.of(modifiedChpInput))
-        .evCS(Set.of(modifiedEvCSInput))
+        .evcs(Set.of(modifiedEvCSInput))
         .evs(Set.of(modifiedEvInput))
         .fixedFeedIn(Set.of(modifiedFixedFeedInInput))
         .heatPumps(Set.of(modifiedHpInput))
@@ -73,13 +70,12 @@ class SystemParticipantsTest extends Specification {
         .pvPlants(Set.of(modifiedPvInput))
         .storages(Set.of(modifiedStorageInput))
         .wecPlants(Set.of(modifiedWecInput))
-        .emSystems(Set.of(modifiedEmInput))
         .build()
 
     then:
     modifiedSystemParticipants.bmPlants.first() == modifiedBmInput
     modifiedSystemParticipants.chpPlants.first() == modifiedChpInput
-    modifiedSystemParticipants.evCS.first() == modifiedEvCSInput
+    modifiedSystemParticipants.evcs.first() == modifiedEvCSInput
     modifiedSystemParticipants.evs.first() == modifiedEvInput
     modifiedSystemParticipants.fixedFeedIns.first() == modifiedFixedFeedInInput
     modifiedSystemParticipants.heatPumps.first() == modifiedHpInput
@@ -87,6 +83,5 @@ class SystemParticipantsTest extends Specification {
     modifiedSystemParticipants.pvPlants.first() == modifiedPvInput
     modifiedSystemParticipants.storages.first() == modifiedStorageInput
     modifiedSystemParticipants.wecPlants.first() == modifiedWecInput
-    modifiedSystemParticipants.emSystems.first() == modifiedEmInput
   }
 }